Move kohana config to a new config helper.
